Firebase and Android setup
