--- hooks/usePushNotificationHandler.android.tsx
@@ -1,1 +1,0 @@
-export default function usePushNotificationHandler() {}
